refactor: run tests with robolectric
Note: Not fully working. Some tests passing, some not. Get tests passing is the next step.
