+++ app/build.gradle
@@ -108,12 +108,14 @@ android {
         }
     }
     sourceSets {
-        androidTest.assets.srcDirs += files("$projectDir/schemas".toString())
+        test {
+            java.srcDir 'src/sharedTest/java'
+        }
+        androidTest {
+            java.srcDir 'src/sharedTest/java'
 
-        // Put code into below directory that you want shared between unit and integration tests.
-        String sharedTestDir = 'src/sharedTest/java'
-        test.java.srcDirs += sharedTestDir
-        androidTest.java.srcDirs += sharedTestDir
+            assets.srcDirs += files("$projectDir/schemas".toString())
+        }
     }
 }
 
@@ -125,7 +127,7 @@ dependencies {
     implementation "androidx.appcompat:appcompat:1.1.0"
     implementation "androidx.fragment:fragment-ktx:1.2.0"
     debugImplementation 'androidx.fragment:fragment-testing:1.2.0'
-    implementation "com.google.android.material:material:1.0.0"
+    implementation "com.google.android.material:material:1.1.0"
     implementation "androidx.recyclerview:recyclerview:1.1.0"
     implementation "androidx.preference:preference:1.1.0"
     // Required to support SDK < 21
@@ -218,6 +220,11 @@ dependencies {
 
     // Android Debug Database
     debugImplementation 'com.amitshekhar.android:debug-db:1.0.4'
+
+    // Robolectric
+    def robolectricVersion = "4.3.1"
+    testImplementation "org.robolectric:robolectric:$robolectricVersion"
+    androidTestImplementation "org.robolectric:annotations:$robolectricVersion"
 
     // JUnit
     def junitVersion = "4.12"
