+++ app/src/main/java/com/levibostian/androidblanky/view/ui/MainApplication.kt
@@ -5,10 +5,6 @@ import android.app.Application
 import android.app.Service
 import android.content.Context
 import androidx.fragment.app.Fragment
-import androidx.work.Constraints
-import androidx.work.NetworkType
-import androidx.work.PeriodicWorkRequest
-import androidx.work.WorkManager
 import com.crashlytics.android.Crashlytics
 import com.crashlytics.android.core.CrashlyticsCore
 import com.levibostian.androidblanky.service.pendingtasks.PendingTasksFactory
@@ -20,12 +16,13 @@ import com.levibostian.wendy.service.Wendy
 import io.fabric.sdk.android.Fabric
 import java.util.concurrent.TimeUnit
 import androidx.multidex.MultiDex
+import androidx.work.*
 import com.levibostian.androidblanky.BuildConfig
 import com.levibostian.androidblanky.di.*
 import javax.inject.Inject
 
 @OpenForTesting
-class MainApplication: Application() {
+class MainApplication: Application(), Configuration.Provider {
 
     lateinit var appComponent: AppGraph
 
@@ -71,5 +68,7 @@ class MainApplication: Application() {
                 .build()
         WorkManager.getInstance(this).enqueue(pendingTaskWorkerBuilder.setConstraints(constraints).build())
     }
+
+    override fun getWorkManagerConfiguration(): Configuration = Configuration.Builder().build()
 
 }
