+++ app/src/sharedTest/java/com/levibostian/androidblanky/rule/RxImmediateSchedulerRule.kt
@@ -1,4 +1,4 @@
-package com.levibostian.androidblanky
+package com.levibostian.androidblanky.rule
 
 import io.reactivex.Scheduler
 import io.reactivex.android.plugins.RxAndroidPlugins
